Move new methods used in the PluginManager into their own utility class. Make them static. Consolidate all tests into an appropriately named test class.
